% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{featureData}
\alias{featureData}
\alias{featureData,Assignment-method}
\alias{correlations}
\alias{correlations,Assignment-method}
\alias{relationships}
\alias{relationships,Assignment-method}
\alias{iterations}
\alias{iterations,Assignment-method}
\alias{graph}
\alias{graph,Assignment-method}
\alias{components}
\alias{components,Assignment-method}
\alias{featureComponents}
\alias{featureComponents,Assignment-method}
\alias{component}
\alias{component,Assignment-method}
\alias{assignments}
\alias{assignments,Assignment-method}
\alias{assignedData}
\alias{assignedData,Assignment-method}
\alias{summariseAssignments}
\alias{summariseAssignments,Assignment-method}
\title{Assignment accessors}
\usage{
featureData(assignment)

\S4method{featureData}{Assignment}(assignment)

correlations(assignment)

\S4method{correlations}{Assignment}(assignment)

relationships(assignment)

\S4method{relationships}{Assignment}(assignment)

iterations(assignment)

\S4method{iterations}{Assignment}(assignment)

graph(assignment, iteration, type = c("selected", "all"))

\S4method{graph}{Assignment}(assignment, iteration, type = c("selected", "all"))

components(assignment, iteration, type = c("selected", "all"))

\S4method{components}{Assignment}(assignment, iteration, type = c("selected", "all"))

featureComponents(assignment, feature, type = c("selected", "all"))

\S4method{featureComponents}{Assignment}(assignment, feature, type = c("selected", "all"))

component(assignment, component, iteration, type = c("selected", "all"))

\S4method{component}{Assignment}(assignment, component, iteration, type = c("selected", "all"))

assignments(assignment)

\S4method{assignments}{Assignment}(assignment)

assignedData(assignment)

\S4method{assignedData}{Assignment}(assignment)

summariseAssignments(assignment)

\S4method{summariseAssignments}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}

\item{iteration}{the assignment iteration}

\item{type}{the graph type to return. \code{filtered} returns the assignment graph after component selection. \code{all} returns all assignment components.}

\item{feature}{feature information to extract}

\item{component}{component number to extract}
}
\value{
A tibble or \code{tbl_graph} containing assignment results depending on the method used.
}
\description{
Access methods for \code{Assignment} S4 class
}
\details{
\itemize{
\item \code{featureData} - Return the initially specifed \emph{m/z} feature data.
\item \code{correlations} - Return the correlation analysis results.
\item \code{relationships} - Return the calculated relationships.
\item \code{iterations} - Return the assignment iteration performed.
\item \code{graph} - Return a selected graph.
\item \code{components} - Return the component information for an assignment iteration.
\item \code{featureComponents} - Return the component information for a selected feature.
\item \code{component} - Extract a component graph.
\item \code{assignments} - Return the molecular formulas assigned to the \emph{m/z} features.
\item \code{assignedData} - Return the \emph{m/z} peak intensity matrix with the molecular formula assignments included in the column names.
\item \code{summariseAssignments} - Return a tibble of the assignments summarised by molecular formula.
}
}
\examples{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

## Return feature data
featureData(mf_assignments)

## Return correlations
correlations(mf_assignments)

## Return relationships
relationships(mf_assignments)

## Return the available iterations
iterations(mf_assignments)

## Return a selected graph
graph(mf_assignments,'A&I1')

## Return a component information for a selected graph
components(mf_assignments,'A&I1')

## Return a component information for a selected feature
featureComponents(mf_assignments,'n191.01962')

 ## Extract a component graph
component(mf_assignments,1,'A&I1')

## Return assignments
assignments(mf_assignments)

## Return an m/z intensity matrix with the assignments included
## in the column names
assignedData(mf_assignments)

## Return the assignments summarised by molecular formula
summariseAssignments(mf_assignments)
}
