% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignedData}
\alias{assignedData}
\alias{assignedData,Assignment-method}
\title{assignedData}
\usage{
assignedData(assignment)

\S4method{assignedData}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\value{
A tibble containing the original feature data with molecular formula assignments added to teh column names.
}
\description{
Return data table used for assignments with feature assignments added to column names.
}
\examples{
\dontrun{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

assignedData(mf_assignments)
}
}
