% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\docType{class}
\name{Assignment-class}
\alias{Assignment-class}
\title{Assignment}
\description{
An S4 class to store molecular formula assignment results.
}
\section{Slots}{

\describe{
\item{\code{log}}{a list containing assignment logs}

\item{\code{data}}{a tibble containing the \emph{m/z} peak intensity matrix}

\item{\code{correlations}}{a tibble containing the correlations analysis results}

\item{\code{relationships}}{a tibble containing the calculated mathematical relationships}

\item{\code{addIsoAssign}}{a list containing the results of the adduct and isotope assignment iterations}

\item{\code{transAssign}}{a list containing the results of the transformation assignment iterationst}

\item{\code{assignments}}{a tibble containing the assigned molecular formulas}
}}

