% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R, R/addIsoAssign.R,
%   R/transformationAssign.R, R/relationships.R
\name{calcCorrelations}
\alias{calcCorrelations}
\alias{calcCorrelations,Assignment-method}
\alias{addIsoAssign}
\alias{addIsoAssign,Assignment-method}
\alias{transformationAssign}
\alias{transformationAssign,Assignment-method}
\alias{calcRelationships}
\alias{calcRelationships,Assignment-method}
\title{Molecular formula assignment methods}
\usage{
calcCorrelations(assignment)

\S4method{calcCorrelations}{Assignment}(assignment)

addIsoAssign(assignment)

\S4method{addIsoAssign}{Assignment}(assignment)

transformationAssign(assignment)

\S4method{transformationAssign}{Assignment}(assignment)

calcRelationships(assignment)

\S4method{calcRelationships}{Assignment}(assignment)
}
\arguments{
\item{assignment}{an object of S4 class \code{Assignment}}
}
\value{
An object of S4 class \code{Assignment} containing molecular formula assignments.
}
\description{
These methods provide the access to performing the individual steps of the molecular
formula assignment approach. See Details for more information of when it is best to use these
instead of \code{assignMFs()}.
}
\details{
In circumstances where the molecular formula assignment approach has high memory requirements,
such as where there are many correlations (> 2 million) or many high \emph{m/z} (>700), it may be
preferable to perform the assignment steps separately as opposed to using \code{assignMFs()}. This
can reduce the memory overheads required to successfully assign molecular formulas to the data
and also enable the possibility of objects to be saved and/or unloaded between the assignment
steps where needed.
}
\examples{
\dontrun{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignment(feature_data,p)

mf_assignments <- mf_assignments \%>\% 
   calcCorrelations() \%>\% 
   calcRelationships() \%>\% 
   addIsoAssign() \%>\% 
   transformationAssign()
}
}
