% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access-methods.R
\docType{methods}
\name{assignmentData}
\alias{assignmentData}
\alias{assignmentData,Assignment-method}
\title{assignmentData}
\usage{
assignmentData(assignment)

\S4method{assignmentData}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Return data table used for assignments.
}
