% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{class}
\name{AssignmentParameters-class}
\alias{AssignmentParameters-class}
\title{S4 class for assignment parameters}
\description{
An S4 class to store assignment parameters.
}
\section{Slots}{

\describe{
\item{\code{technique}}{the analytical technique}

\item{\code{correlations_parameters}}{a list of correlation parameters to be passed to \code{metabolyseR::correlations()}}

\item{\code{max_M}}{the maximum molecular mass for which to assign molecular formulas}

\item{\code{MF_rank_threshold}}{rank threshold for molecular formula selection}

\item{\code{ppm}}{the parts per million error threshold}

\item{\code{limit}}{the atomic mass unit deviation limit for relationship calculation}

\item{\code{RT_diff_limit}}{the limit for retention time differences for correlated features in adduct and isotopic assignment}

\item{\code{adducts}}{a list of character vectors containing the adducts names. List element names should denote ionisation mode. The order that these adducts are provided denotes their expected relative importance to assignments with the first expected to be the most common and the last the least common within each ionisation mode.}

\item{\code{isotopes}}{a character vector of isotopes to use. Similarly to the adducts, their order denotes the expected commonality in the data.}

\item{\code{transformations}}{a character vector of transformations molecular formula changes}

\item{\code{adduct_rules}}{a tibble containing the adduct formation rules as returned by \code{mzAnnotation::adduct_rules()}}

\item{\code{isotope_rules}}{a tibble containing the isotope rules as returned by \code{mzAnnotation::isotope_rules()}}

\item{\code{transformation_rules}}{tibble containing the transformation rules as returned by \code{mzAnnotation::transformation_rules()}}
}}

