% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignmentParameters.R
\name{assignmentParameters}
\alias{assignmentParameters}
\title{annotationParameters}
\usage{
assignmentParameters(technique = NULL)
}
\arguments{
\item{technique}{technique to use for assignment. \code{NULL} prints available techniques}
}
\description{
Return assignment parameters for a specified technique.
}
