% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access-methods.R
\docType{methods}
\name{assignments}
\alias{assignments}
\alias{assignments,Assignment-method}
\title{assignments-Assignment}
\usage{
assignments(assignment)

\S4method{assignments}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Get table of assigned features from an Assignment
}
