% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{availableTechniques}
\alias{availableTechniques}
\title{Available analytical techniques}
\usage{
availableTechniques()
}
\value{
A \code{character} vector of technique names.
}
\description{
The available analytical techniques for molecular formula assignment parameters.
}
\examples{
availableTechniques()
}
