% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continueAssignment.R
\name{continueAssignment}
\alias{continueAssignment}
\title{continueAssignment}
\usage{
continueAssignment(assignment)
}
\arguments{
\item{assignment}{an S4 object of class Assignment}
}
\description{
continue a failed assignment
}
