% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\docType{data}
\name{correlations}
\alias{correlations}
\title{correlations}
\format{A tibble containining 60177 rows and 4 variablesL
\describe{
  \item{Feature1}{first correlated feature}
  \item{Feature2}{second correlated feature}
  \item{log2IntensityRatio}{log2 intensity ratio between the features}
  \item{r}{correlation coefficient between the features}
}}
\usage{
correlations
}
\description{
example correlations table computed using the correlations analysis element in the metabolyseR package, from the example FIE-HRMS B. distachyon ecotype data in the metaboData package.
}
\keyword{datasets}
