% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\docType{data}
\name{feature_data}
\alias{feature_data}
\title{Example feature data}
\format{
A tibble containing 60 rows and 10 variables.
}
\usage{
feature_data
}
\description{
An example \code{m/z} peak intensity matrix containing total ion count normalised positive and negative mode flow infusion electrospray ionisation mass spectrometry \emph{m/z} features.
}
\keyword{datasets}
