% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{nodes}
\alias{nodes}
\alias{edges}
\title{Extract graph attributes}
\usage{
nodes(graph)

edges(graph)
}
\arguments{
\item{graph}{object of class tbl_graph}
}
\description{
Extract node or edge attributes from a \emph{tidygraph} \code{tbl_graph} object.
}
\examples{
a_graph <- tidygraph::tbl_graph(
  nodes = data.frame(
    name = c('a','b','c')
  ), 
  edges = data.frame(
    from = c(1,2),
    to = c(2,3),
    type = c(1,2)
  ))

## Extract graph nodes
nodes(a_graph)

## Extract graph edges
edges(a_graph)
}
