% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{iso}
\alias{iso}
\alias{iso,AssignmentParameters-method}
\alias{iso<-}
\alias{iso<-,AssignmentParameters-method}
\alias{add}
\alias{add,AssignmentParameters-method}
\alias{add<-}
\alias{add<-,AssignmentParameters-method}
\title{Parameter get and set methods}
\usage{
iso(x)

\S4method{iso}{AssignmentParameters}(x)

iso(x) <- value

\S4method{iso}{AssignmentParameters}(x) <- value

add(x)

\S4method{add}{AssignmentParameters}(x)

add(x) <- value

\S4method{add}{AssignmentParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{AssignmentParameters}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{AssignmentParameters} S4 class.
}
\examples{
assignment_parameters <- assignmentParameters('FIE')

## Return isotopes
iso(assignment_parameters)

## Set isotopes
iso(assignment_parameters) <- '13C'

## Return adducts
add(assignment_parameters)

## Set adducts
add(assignment_parameters) <- list(n = c('[M-H]1-','[M+Cl]1-'),
                                   p = c('[M+H]1+','[M+K]1+'))

}
