% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakData.R
\docType{data}
\name{peakData}
\alias{peakData}
\title{peakData}
\format{
A tibble containining 60 rows and 1003 variables
}
\usage{
peakData
}
\description{
example peak intensity table of sample data from the example FIE-HRMS B. distachyon ecotype data in the metaboData package.
}
\keyword{datasets}
