% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdductDist-method.R
\docType{methods}
\name{plotAdductDist,Assignment-method}
\alias{plotAdductDist,Assignment-method}
\title{plotAdductDist-Assignment}
\usage{
\S4method{plotAdductDist}{Assignment}(assignment)
}
\description{
plotAdductDist-Assignment
}
