% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdductDist.R
\name{plotAdductDist}
\alias{plotAdductDist}
\alias{plotAdductDist,Assignment-method}
\title{Plot adduct frequency distributions}
\usage{
plotAdductDist(assignment)

\S4method{plotAdductDist}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Plot adduct frequency distributions.
}
