% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plotComponent}
\alias{plotComponent}
\alias{plotComponent,Assignment-method}
\title{Plot a component}
\usage{
plotComponent(
  assignment,
  component,
  iteration,
  type = c("selected", "all"),
  label_size = 3,
  axis_offset = 0.1,
  border = NA,
  highlight = NA
)

\S4method{plotComponent}{Assignment}(
  assignment,
  component,
  iteration,
  type = c("selected", "all"),
  label_size = 3,
  axis_offset = 0.1,
  border = NA,
  highlight = NA
)
}
\arguments{
\item{assignment}{S4 object of class Assignment}

\item{component}{component number to extract}

\item{iteration}{the assignment iteration}

\item{type}{the graph type to return. \code{selected} returns the assignment graph after component selection. \code{all} returns all assignment components.}

\item{label_size}{node label size}

\item{axis_offset}{axis proportion by which to increase axis limits. Prevents cut off of node labels.}

\item{border}{specify a plot border colour}

\item{highlight}{specify a feature node to highlight}
}
\description{
Plot a molecular formula component graph.
}
\examples{
\dontrun{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

plotComponent(mf_assignments,1,'A&I1')
}
}
