% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plotFeatureComponents}
\alias{plotFeatureComponents}
\alias{plotFeatureComponents,Assignment-method}
\title{Plot the solutions for a feature}
\usage{
plotFeatureComponents(
  assignment,
  feature,
  iteration,
  type = c("all", "selected"),
  max_components = 6,
  label_size = 3,
  axis_offset = 0.1
)

\S4method{plotFeatureComponents}{Assignment}(
  assignment,
  feature,
  iteration,
  type = c("all", "selected"),
  max_components = 6,
  label_size = 2,
  axis_offset = 0.05
)
}
\arguments{
\item{assignment}{S4 object of class Assignent}

\item{feature}{name of feature to plot}

\item{iteration}{components from which iteration to plot}

\item{type}{the graph type to return. \code{all} returns all assignment components. \code{selected} returns the assignment graph after component selection.}

\item{max_components}{maximum number of components to plot}

\item{label_size}{node label size}

\item{axis_offset}{axis proportion by which to increase axis limits. Prevents cut off of node labels.}
}
\description{
Plot possible MF solutions for a given feature.
}
