% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotNetwork-method.R
\docType{methods}
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,Assignment-method}
\title{plotNetwork-Assignment}
\usage{
plotNetwork(assignment, layout = "kk", rThreshold = 0.7, labels = F)

\S4method{plotNetwork}{Assignment}(assignment, layout = "kk",
  rThreshold = 0.7, labels = F)
}
\description{
plotNetwork-Assignment
}
