% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotSpectrum-method.R
\docType{methods}
\name{plotSpectrum}
\alias{plotSpectrum}
\alias{plotSpectrum,Assignment-method}
\title{plotSpectrum}
\usage{
plotSpectrum(assignment, MF)

\S4method{plotSpectrum}{Assignment}(assignment, MF)
}
\arguments{
\item{assignment}{S4 object of class Assignment}

\item{MF}{molecular formula}
}
\description{
Plot a spectrum for a given molecular formula
}
