% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdductDist.R, R/plot_components.R,
%   R/plotSpectrum.R
\name{plotAdductDist}
\alias{plotAdductDist}
\alias{plotAdductDist,Assignment-method}
\alias{plotComponent}
\alias{plotComponent,Assignment-method}
\alias{plotFeatureComponents}
\alias{plotFeatureComponents,Assignment-method}
\alias{plotSpectrum}
\alias{plotSpectrum,Assignment-method}
\title{Plot assignment results}
\usage{
plotAdductDist(assignment)

\S4method{plotAdductDist}{Assignment}(assignment)

plotComponent(
  assignment,
  component,
  iteration,
  type = c("selected", "all"),
  label_size = 3,
  axis_offset = 0.1,
  border = NA,
  highlight = NA
)

\S4method{plotComponent}{Assignment}(
  assignment,
  component,
  iteration,
  type = c("selected", "all"),
  label_size = 3,
  axis_offset = 0.1,
  border = NA,
  highlight = NA
)

plotFeatureComponents(
  assignment,
  feature,
  iteration,
  type = c("all", "selected"),
  max_components = 6,
  label_size = 3,
  axis_offset = 0.1
)

\S4method{plotFeatureComponents}{Assignment}(
  assignment,
  feature,
  iteration,
  type = c("all", "selected"),
  max_components = 6,
  label_size = 2,
  axis_offset = 0.05
)

plotSpectrum(assignment, MF)

\S4method{plotSpectrum}{Assignment}(assignment, MF)
}
\arguments{
\item{assignment}{an object of S4 class Assignment}

\item{component}{component number to plot}

\item{iteration}{the assignment iteration of the component or components}

\item{type}{the graph type to return. \code{selected} returns the assignment graph after component selection. \code{all} returns all assignment components.}

\item{label_size}{node label size}

\item{axis_offset}{axis proportion by which to increase axis limits. Prevents cut off of node labels.}

\item{border}{specify a plot border colour}

\item{highlight}{specify a feature node to highlight}

\item{feature}{the \emph{m/z} feature to plot}

\item{max_components}{themaximum number of components to plot}

\item{MF}{the assigned molecular formula to plot}
}
\description{
Plot molecular formula assignment results.
}
\details{
\itemize{
\item \code{plotComponent} - Plot a molecular formula component graph.
\item \code{plotFeatureComponents} - Plot the possible component graphs for a given feature.
\item \code{plotAdductDist} - Plot frequency distributions of the assigned adducts.
\item \code{plotSpectrum} - Plot the spectrum of an assigned molecular formula.
}
}
\examples{
library(ggraph)
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

## Plot a component
plotComponent(mf_assignments,1,'A&I1')

## Plot the components for a feature
plotFeatureComponents(mf_assignments,'n191.01962','A&I1')

## Plot the adduct distributions
plotAdductDist(mf_assignments)

## Plot the spectrum of an assigned molecular formula
plotSpectrum(mf_assignments,'C6H8O7')
}
