% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,Assignment-method}
\alias{show,Assignment-method}
\title{show-Assignment}
\usage{
\S4method{show}{Assignment}(object)
}
\arguments{
\item{object}{S4 object of class Assignment}
}
\description{
show mehtod for Assignment class.
}
