% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/summariseAssignment-method.R
\docType{methods}
\name{summariseAssignment}
\alias{summariseAssignment}
\alias{summariseAssignment,Assignment-method}
\title{summariseAssignment-Assignment}
\usage{
summariseAssignment(assignment)

\S4method{summariseAssignment}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Summarise features assigned to moleuclar formulas.
}
