% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{summariseAssignments}
\alias{summariseAssignments}
\alias{summariseAssignments,Assignment-method}
\title{Summarise assignments}
\usage{
summariseAssignments(assignment)

\S4method{summariseAssignments}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\value{
A tibble containing the feature assignments summarised by molecular formula.
}
\description{
Summarise features assigned to molecular formulas.
}
\examples{
\dontrun{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

summariseAssignments(mf_assignments)
}
}
