% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{relationships}
\alias{relationships}
\alias{relationships,Assignment-method}
\alias{assignments}
\alias{assignments,Assignment-method}
\title{Assignment accessors}
\usage{
relationships(assignment)

\S4method{relationships}{Assignment}(assignment)

assignments(assignment)

\S4method{assignments}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Access methods for Assignment S4 class
}
