% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{featureData}
\alias{featureData}
\alias{featureData,Assignment-method}
\alias{correlations}
\alias{correlations,Assignment-method}
\alias{relationships}
\alias{relationships,Assignment-method}
\alias{iterations}
\alias{iterations,Assignment-method}
\alias{graph}
\alias{graph,Assignment-method}
\alias{assignments}
\alias{assignments,Assignment-method}
\title{Assignment accessors}
\usage{
featureData(assignment)

\S4method{featureData}{Assignment}(assignment)

correlations(assignment)

\S4method{correlations}{Assignment}(assignment)

relationships(assignment)

\S4method{relationships}{Assignment}(assignment)

iterations(assignment)

\S4method{iterations}{Assignment}(assignment)

graph(assignment, iteration, type = c("filtered", "all"))

\S4method{graph}{Assignment}(assignment, iteration, type = c("filtered", "all"))

assignments(assignment)

\S4method{assignments}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}

\item{iteration}{the assignment iteration}

\item{type}{the graph type to return. \code{filtered} returns the assignment graph after component selection. \code{all} returns all assignment components.}
}
\description{
Access methods for Assignment S4 class
}
\examples{
assignment <- new('Assignment',
                  data = feature_data)

## Return feature data
featureData(assignment)

## Return correlations
correlations(assignment)

## Return relationships
relationships(assignment)

## Return the available iterations
iterations(assignment)

## Return a selected graph
\dontrun{
graph(assignment,'A&I1')
}

## Return assignments
assignments(assignment)
}
