% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assignMFs}
\alias{assignMFs}
\alias{continueAssignment}
\title{Assign molecular formulas}
\usage{
assignMFs(dat, parameters, verbose = TRUE)

continueAssignment(assignment)
}
\arguments{
\item{dat}{tibble containing the peak intensities of m/z for which to assign molecular formulas}

\item{parameters}{an S4 object of class AssignmentParamters containing the parameters for molecular formula assignment}

\item{verbose}{should output be printed to the console}
}
\description{
assign molecular formulas to a set of given m/z.
}
\examples{
plan(future::sequential)
p <- assignmentParameters('FIE')

assignment <- assignMFs(peakData,p)

}
