% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assignMFs}
\alias{assignMFs}
\alias{assignMFs,tbl_df-method}
\alias{assignMFs,AnalysisData-method}
\alias{assignMFs,Analysis-method}
\title{Assign molecular formulas}
\usage{
assignMFs(
  feature_data,
  parameters = assignmentParameters("FIE-HRMS"),
  verbose = TRUE,
  ...
)

\S4method{assignMFs}{tbl_df}(
  feature_data,
  parameters = assignmentParameters("FIE-HRMS"),
  verbose = TRUE
)

\S4method{assignMFs}{AnalysisData}(
  feature_data,
  parameters = assignmentParameters("FIE"),
  verbose = TRUE
)

\S4method{assignMFs}{Analysis}(
  feature_data,
  parameters = assignmentParameters("FIE"),
  verbose = TRUE,
  type = c("pre-treated", "raw")
)
}
\arguments{
\item{feature_data}{a tibble or an object of S4 class \code{AnalysisData} or \code{Analysis} containing the feature intensity matrix of m/z for which to assign molecular formulas. See details.}

\item{parameters}{an S4 object of class \code{AssignmentParamters} containing the parameters for molecular formula assignment}

\item{verbose}{should progress output be printed to the console}

\item{...}{arguments to pass to the relevant method}

\item{type}{\code{raw} or \code{pre-treated} data on which to perform assignment when argument \code{feature_data} is of class \code{Analysis}}
}
\description{
assign molecular formulas to a set of given m/z.
}
\details{
If argument \code{feature_data} is specified as a tibble, this should be a feature intensity matrix where the columns are the \code{m/z} features to assign and the rows are the individual observations, with the cells as abundance values.
}
\examples{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)

}
