% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assignMFs}
\alias{assignMFs}
\alias{assignMFs,tbl_df-method}
\alias{assignMFs,AnalysisData-method}
\alias{assignMFs,Analysis-method}
\title{Perform molecular formula assignment}
\usage{
assignMFs(
  feature_data,
  parameters = assignmentParameters("FIE-HRMS"),
  verbose = TRUE,
  ...
)

\S4method{assignMFs}{tbl_df}(
  feature_data,
  parameters = assignmentParameters("FIE-HRMS"),
  verbose = TRUE
)

\S4method{assignMFs}{AnalysisData}(
  feature_data,
  parameters = assignmentParameters("FIE"),
  verbose = TRUE
)

\S4method{assignMFs}{Analysis}(
  feature_data,
  parameters = assignmentParameters("FIE"),
  verbose = TRUE,
  type = c("pre-treated", "raw")
)
}
\arguments{
\item{feature_data}{a tibble or an object of S4 class \code{AnalysisData} or \code{Analysis} containing the feature intensity matrix of m/z for which to assign molecular formulas. See details.}

\item{parameters}{an S4 object of class \code{AssignmentParamters} containing the parameters for molecular formula assignment}

\item{verbose}{should progress output be printed to the console}

\item{...}{arguments to pass to the relevant method}

\item{type}{\code{pre-treated} or \code{raw} data on which to perform assignment when argument \code{feature_data} is of S4 class \code{Analysis}}
}
\description{
Perform automated molecular formula assignment.
}
\details{
If argument \code{feature_data} is specified as a tibble, this should be a feature intensity matrix where
the columns are the \code{m/z} features to assign and the rows are the individual observations, with the
cells as abundance values. he m/z features provided as column names should be in the form of
<ionisation_mode><m/z>@<retention_time>. Ionisation mode should be given as a prefix n or p for negative
or positive ionisation modes respectively. Feature m/z should be provided to an accuracy of least 5 decimal
places. The retention time portion (@<retention_time>) is only required for LC-MS data and should be provided
in minutes.
}
\examples{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

assignments <- assignMFs(feature_data,p)

}
