% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignedData}
\alias{assignedData}
\alias{assignedData,Assignment-method}
\title{assignedData}
\usage{
assignedData(assignment)

\S4method{assignedData}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Return data table used for assignments with feature assignments added to column names.
}
