% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Assignment-class}
\alias{Assignment-class}
\title{Assignment}
\description{
An S4 class to store assignment results
}
\section{Slots}{

\describe{
\item{\code{parameters}}{An S4 object of class AssignmentParameters containing the assignment parameters}

\item{\code{correlations}}{A tibble containing the correlations}

\item{\code{relationships}}{A tibble containing the predicted relationships}

\item{\code{addIsoAssign}}{A list containing the results of the adduct and isotope assignment}

\item{\code{transAssign}}{A list containing the results of the transformation assignment}

\item{\code{assignments}}{A tibble containing the assigned molecular formulas}
}}

