% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\docType{class}
\name{Assignment-class}
\alias{Assignment-class}
\title{Assignment}
\description{
An S4 class to store assignment results
}
\section{Slots}{

\describe{
\item{\code{log}}{list containing assignment logs}

\item{\code{flags}}{charactor vector containing completed assignment elements}

\item{\code{data}}{A tibble containing the peak intensity matrix}

\item{\code{correlations}}{A tibble containing the correlations}

\item{\code{preparedCorrelations}}{A tibble containing the prepared correlations ready for analysis}

\item{\code{relationships}}{A tibble containing the predicted relationships}

\item{\code{addIsoAssign}}{A list containing the results of the adduct and isotope assignment}

\item{\code{transAssign}}{A list containing the results of the transformation assignment}

\item{\code{assignments}}{A tibble containing the assigned molecular formulas}
}}

