% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{AssignmentParameters-class}
\alias{AssignmentParameters-class}
\title{AssignmentParameters}
\description{
An S4 class to store assignment parameters.
}
\section{Slots}{

\describe{
\item{\code{technique}}{assignment technique to use}

\item{\code{maxM}}{maximum M for which to assign molecular formulas}

\item{\code{maxMFscore}}{threshold for molecular formula score}

\item{\code{ppm}}{ppm threshold}

\item{\code{limit}}{amu deviation limit for relationship prediction}

\item{\code{isotopes}}{character vector of isotopes to use}

\item{\code{adducts}}{named list of character vectors containing the adducuts to use for each mode}

\item{\code{nCores}}{number of cores to use for parallisation}

\item{\code{clusterType}}{cluster type to use for parallisation}
}}

