% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{class}
\name{AssignmentParameters-class}
\alias{AssignmentParameters-class}
\title{S4 class for assignment parameters}
\description{
An S4 class to store assignment parameters.
}
\section{Slots}{

\describe{
\item{\code{technique}}{assignment technique to use}

\item{\code{correlations_parameters}}{list of correlation parameters to be passed to \code{metabolyseR::correlation()}}

\item{\code{max_M}}{maximum M for which to assign molecular formulas}

\item{\code{MF_rank_threshold}}{rank threshold for molecular formula selection}

\item{\code{ppm}}{ppm threshold}

\item{\code{limit}}{amu deviation limit for relationship prediction}

\item{\code{RT_diff_limit}}{limit for retention time differences for correlated features in adduct and isotopic assignment}

\item{\code{adducts}}{list of character vectors containing the adducts to use. List element names should denote ionisation mode. The order that these adducts are provided denotes their expected relative importance to assignments with the first expected to be the most common and the last the least common within each ionisation mode.}

\item{\code{isotopes}}{character vector of isotopes to use. Similarly to the adducts, their order denotes the expected commonality in the data.}

\item{\code{transformations}}{character vector of transformations to use}

\item{\code{adduct_rules}}{tibble containing adduct formation rules as returned by \code{mzAnnotation::adducts()}}

\item{\code{isotope_rules}}{tibble containing isotope rules as returned by \code{mzAnnotation::isotopes()}}

\item{\code{transformation_rules}}{tibble containing transformation rules as returned by \code{mzAnnotation::transformations()}}
}}

