% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{assignmentParameters}
\alias{assignmentParameters}
\title{Assignment parameters}
\usage{
assignmentParameters(technique = availableTechniques())
}
\arguments{
\item{technique}{technique to use for assignment}
}
\description{
Return default assignment parameters for a given technique.
}
