% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{assignmentParameters}
\alias{assignmentParameters}
\title{Assignment parameters}
\usage{
assignmentParameters(technique = availableTechniques())
}
\arguments{
\item{technique}{technique to use for assignment}
}
\value{
An object of S4 class \code{AssignmentParameters}
}
\description{
Return the default molecular formula assignment parameters for a given analytical technique.
}
\examples{
assignmentParameters('FIE-HRMS')
}
