% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMFs.R
\name{assignMFs}
\alias{assignMFs}
\title{assignMFs}
\usage{
assignMFs(correlations, parameters)
}
\arguments{
\item{correlations}{table containing correlations of m/z to assign molecular formulas}

\item{parameters}{an S4 object of class AssignmentParamters containing the parameters for molecular formula assignment}
}
\description{
assign molecular formulas to a set of given m/z.
}
\examples{
\dontrun{
res <- assignMFs(correlations[20000:21000,],assignmentParameters('FIE'))
}
}
