% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMFs.R
\name{assignMFs}
\alias{assignMFs}
\title{assignMFs}
\usage{
assignMFs(dat, parameters, verbose = T)
}
\arguments{
\item{dat}{tibble containing the peak intensities of m/z for which to assign molecular formulas}

\item{parameters}{an S4 object of class AssignmentParamters containing the parameters for molecular formula assignment}

\item{verbose}{should output be printed to the console}
}
\description{
assign molecular formulas to a set of given m/z.
}
\examples{
p <- assignmentParameters('FIE')
p@nCores <- 2

dat <- dplyr::select(peakData,n191.01962,n192.02306,n193.02388,n226.99693,n228.97636,n228.99274,n231.00069,n384.0495,n385.04874)

assignment <- assignMFs(dat,p)

}
