% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\docType{data}
\name{feature_data}
\alias{feature_data}
\title{Example feature data}
\format{
A tibble containing 60 rows and 1003 variables.
}
\usage{
feature_data
}
\description{
Example mass spectral feature intensity table.
}
\keyword{datasets}
