% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{isotopes}
\alias{isotopes}
\alias{isotopes,AssignmentParameters-method}
\alias{isotopes<-}
\alias{isotopes<-,AssignmentParameters-method}
\alias{adducts}
\alias{adducts,AssignmentParameters-method}
\alias{adducts<-}
\alias{adducts<-,AssignmentParameters-method}
\title{Parameter get and set methods}
\usage{
isotopes(x)

\S4method{isotopes}{AssignmentParameters}(x)

isotopes(x) <- value

\S4method{isotopes}{AssignmentParameters}(x) <- value

adducts(x)

\S4method{adducts}{AssignmentParameters}(x)

adducts(x) <- value

\S4method{adducts}{AssignmentParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{AssignmentParameters}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{AssignmentParameters} S4 class.
}
\examples{
assignment_parameters <- assignmentParameters('FIE')

## Return isotopes
isotopes(assignment_parameters)

## Set isotopes
isotopes(assignment_parameters) <- '13C'

## Return adducts
adducts(assignment_parameters)

## Set adducts
adducts(assignment_parameters) <- list(n = c('[M-H]1-','[M+Cl]1-'),
                                   p = c('[M+H]1+','[M+K]1+'))

}
