% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{iso}
\alias{iso}
\alias{iso,AssignmentParameters-method}
\alias{iso<-}
\alias{iso<-,AssignmentParameters-method}
\title{Parameter get and set methods}
\usage{
iso(x)

\S4method{iso}{AssignmentParameters}(x)

iso(x) <- value

\S4method{iso}{AssignmentParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{AssignmentParameters}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{AssignmentParameters} S4 class.
}
\examples{
assignment_parameters <- assignmentParameters('FIE')

## Return isotopes
iso(assignment_parameters)

}
