% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{technique}
\alias{technique}
\alias{technique,AssignmentParameters-method}
\alias{limit}
\alias{limit,AssignmentParameters-method}
\alias{limit<-}
\alias{limit<-,AssignmentParameters-method}
\alias{maxM}
\alias{maxM,AssignmentParameters-method}
\alias{maxM<-}
\alias{maxM<-,AssignmentParameters-method}
\alias{maxMFscore}
\alias{maxMFscore,AssignmentParameters-method}
\alias{maxMFscore<-}
\alias{maxMFscore<-,AssignmentParameters-method}
\alias{ppm}
\alias{ppm,AssignmentParameters-method}
\alias{ppm<-}
\alias{ppm<-,AssignmentParameters-method}
\alias{isotopes}
\alias{isotopes,AssignmentParameters-method}
\alias{isotopes<-}
\alias{isotopes<-,AssignmentParameters-method}
\alias{adducts}
\alias{adducts,AssignmentParameters-method}
\alias{adducts<-}
\alias{adducts<-,AssignmentParameters-method}
\alias{transformations}
\alias{transformations,AssignmentParameters-method}
\alias{transformations<-}
\alias{transformations<-,AssignmentParameters-method}
\alias{adductRules}
\alias{adductRules,AssignmentParameters-method}
\alias{adductRules<-}
\alias{adductRules<-,AssignmentParameters-method}
\alias{isotopeRules}
\alias{isotopeRules,AssignmentParameters-method}
\alias{isotopeRules<-}
\alias{isotopeRules<-,AssignmentParameters-method}
\alias{transformationRules}
\alias{transformationRules,AssignmentParameters-method}
\alias{transformationRules<-}
\alias{transformationRules<-,AssignmentParameters-method}
\title{Parameter get and set methods}
\usage{
technique(x)

\S4method{technique}{AssignmentParameters}(x)

limit(x)

\S4method{limit}{AssignmentParameters}(x)

limit(x) <- value

\S4method{limit}{AssignmentParameters}(x) <- value

maxM(x)

\S4method{maxM}{AssignmentParameters}(x)

maxM(x) <- value

\S4method{maxM}{AssignmentParameters}(x) <- value

maxMFscore(x)

\S4method{maxMFscore}{AssignmentParameters}(x)

maxMFscore(x) <- value

\S4method{maxMFscore}{AssignmentParameters}(x) <- value

ppm(x)

\S4method{ppm}{AssignmentParameters}(x)

ppm(x) <- value

\S4method{ppm}{AssignmentParameters}(x) <- value

isotopes(x)

\S4method{isotopes}{AssignmentParameters}(x)

isotopes(x) <- value

\S4method{isotopes}{AssignmentParameters}(x) <- value

adducts(x)

\S4method{adducts}{AssignmentParameters}(x)

adducts(x) <- value

\S4method{adducts}{AssignmentParameters}(x) <- value

transformations(x)

\S4method{transformations}{AssignmentParameters}(x)

transformations(x) <- value

\S4method{transformations}{AssignmentParameters}(x) <- value

adductRules(x)

\S4method{adductRules}{AssignmentParameters}(x)

adductRules(x) <- value

\S4method{adductRules}{AssignmentParameters}(x) <- value

isotopeRules(x)

\S4method{isotopeRules}{AssignmentParameters}(x)

isotopeRules(x) <- value

\S4method{isotopeRules}{AssignmentParameters}(x) <- value

transformationRules(x)

\S4method{transformationRules}{AssignmentParameters}(x)

transformationRules(x) <- value

\S4method{transformationRules}{AssignmentParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class \code{AssignmentParameters}}

\item{value}{value to set}
}
\description{
Get and set methods for the \code{AssignmentParameters} S4 class.
}
\examples{
assignment_parameters <- assignmentParameters('FIE')

## Return technique
technique(assignment_parameters)

## Return limit
limit(assignment_parameters)

## Set limit
limit(assignment_parameters) <- 0.002

## Return max M
maxM(assignment_parameters)

## Set max M
maxM(assignment_parameters) <- 500

## Return max MF score
maxMFscore(assignment_parameters)

## Set max MF score
maxMFscore(assignment_parameters) <- 3

## Return ppm
ppm(assignment_parameters)

## Set ppm
ppm(assignment_parameters) <- 3

## Return isotopes
isotopes(assignment_parameters)

## Set isotopes
isotopes(assignment_parameters) <- '13C'

## Return adducts
adducts(assignment_parameters)

## Set adducts
adducts(assignment_parameters) <- list(n = c('[M-H]1-','[M+Cl]1-'),
                                   p = c('[M+H]1+','[M+K]1+'))
                                   
## Return transformations
transformations(assignment_parameters)

## Set transformations
transformations(assignment_parameters) <- "M - [O] + [NH2]"

## Return adduct rules
adductRules(assignment_parameters)

## Set adduct rules
adductRules(assignment_parameters) <- mzAnnotation::adduct_rules())

## Return isotope rules
isotopeRules(assignment_parameters)

## Set isotope rules
isotopeRules(assignment_parameters) <- mzAnnotation::isotope_rules())

## Return transformation rules
transformationRules(assignment_parameters)

## Set transformation rules
transformationRules(assignment_parameters) <- mzAnnotation::transformation_rules())
}
