% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plotComponent}
\alias{plotComponent}
\title{Plot a component}
\usage{
plotComponent(
  assignment,
  component,
  iteration,
  type = c("filtered", "all"),
  label_size = 3,
  axis_offset = 0.1
)
}
\arguments{
\item{assignment}{S4 object of class Assignment}

\item{component}{component number to extract}

\item{iteration}{the assignment iteration}

\item{type}{the graph type to return. \code{filtered} returns the assignment graph after component selection. \code{all} returns all assignment components.}

\item{label_size}{node label size}

\item{axis_offset}{axis proportion by which to increase axis limits. Prevents cut off of node labels.}
}
\description{
Plot a molecular formula component graph.
}
\examples{
\dontrun{
plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

assignment <- assignMFs(feature_data,p)

plotComponent(assignment,1,'A&I1')
}
}
