% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotFeatureSolutions.R
\docType{methods}
\name{plotFeatureSolutions}
\alias{plotFeatureSolutions}
\alias{plotFeatureSolutions,Assignment-method}
\title{plotFeatureSolutions}
\usage{
plotFeatureSolutions(assignment, feature)

\S4method{plotFeatureSolutions}{Assignment}(assignment, feature)
}
\description{
plotFeatureSolutions
}
