% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork-method.R
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,Assignment-method}
\title{Plot an assignment network}
\usage{
plotNetwork(assignment, layout = "stress", rThreshold = 0.7)

\S4method{plotNetwork}{Assignment}(assignment, layout = "stress", rThreshold = 0.7)
}
\arguments{
\item{assignment}{of class Assignment}

\item{layout}{graph layout to use. See \code{\link[ggraph]{ggraph}} for layout options}

\item{rThreshold}{r threhold to use for filtering edge correlation weights}
}
\description{
plot assignment network
}
