% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotNetwork-method.R
\docType{methods}
\name{plotNetwork}
\alias{plotNetwork}
\alias{plotNetwork,Assignment-method}
\title{plotNetwork-Assignment}
\usage{
plotNetwork(assignment, layout = "kk", rThreshold = 0.7, labels = F)

\S4method{plotNetwork}{Assignment}(assignment, layout = "kk",
  rThreshold = 0.7, labels = F)
}
\arguments{
\item{assignment}{of class Assignment}

\item{layout}{graph layout to use. See \code{\link[ggraph]{ggraph}} for layout options}

\item{rThreshold}{r threhold to use for filtering edge correlation weights}

\item{labels}{whether to plot labeld. Defaults to \code{FALSE}}
}
\description{
plot correlation network
}
