% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{summariseAssignment}
\alias{summariseAssignment}
\alias{summariseAssignment,Assignment-method}
\title{Summarise assignments}
\usage{
summariseAssignment(assignment)

\S4method{summariseAssignment}{Assignment}(assignment)
}
\arguments{
\item{assignment}{S4 object of class Assignment}
}
\description{
Summarise features assigned to molecular formulas.
}
